+++ config/initializers/assets.rb
@@ -1,10 +1,10 @@
 # Add Alchemy assets for precompiling
 Rails.application.config.assets.precompile += [
-  'alchemy/alchemy.js',
+  'alchemy/admin/all.css',
+  'alchemy/admin/all.js',
   'alchemy/alchemy-logo.svg',
   'alchemy/favicon.ico',
   'alchemy/preview.js',
-  'alchemy/admin.css',
   'alchemy/menubar.css',
   'alchemy/menubar.js',
   'alchemy/print.css',