+++ app/controllers/alchemy/admin/dashboard_controller.rb
@@ -7,8 +7,8 @@ module Alchemy
       authorize_resource class: :alchemy_admin_dashboard
 
       def index
-        @last_edited_pages = Page.from_current_site.all_last_edited_from(current_alchemy_user)
-        @all_locked_pages = Page.from_current_site.locked
+        @last_edited_pages = Page.all_last_edited_from(current_alchemy_user)
+        @all_locked_pages = Page.locked
         if Alchemy.user_class.respond_to?(:logged_in)
           @online_users = Alchemy.user_class.logged_in.to_a - [current_alchemy_user]
         end