+++ CHANGELOG.md
@@ -35,6 +35,7 @@ __Notable Changes__
   from current site only.
 * Removes the `Language.get_default` method alias for `Language.default`
 * Move site select into pages and languages module to avoid confusion about curent site (#1067)
+* List pages from all sites in currently locked pages tabs and Dashboard widget (#1067)
 
 __Fixed Bugs__