+++ CHANGELOG.md
@@ -4,7 +4,8 @@
 
 * Fix use of Alchemy::Resource with namespaced models (#729)
 * Allow setting the type of EssenceText input fields in the elements.yml via `settings[:input_type]`
-* Admin locale swtiching does not switch language tree any more (#1065)
+* Admin locale switching does not switch language tree any more (#1065)
+* Fixes bug where old site session causes not found errors (#1047)
 
 ## 3.3.0 (2016-05-18)